[attestation-service] E: Refactor API Tests
